+++ WebUI/ServicesModel/PerformedWork/PerformedWorkBaseModel.cs
@@ -6,7 +6,7 @@
 
         public DateTime Date { get; set; } = DateTime.Now;
 
-        public int AmountOfFuel { get; set; }
+        public decimal AmountOfFuel { get; set; }
 
         public decimal FuelPrice { get; set; }
